MenuAbstractModel, addNewUser in Register
