+++ src/main/java/imd/ufrn/br/purposesong/use_case/CreateNewUser.java
@@ -5,6 +5,7 @@ import imd.ufrn.br.purposesong.entity.User;
 
 public class CreateNewUser {
     final UserRepository userRepository;
+
     public CreateNewUser(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
