+++ src/main/resources/imd/ufrn/br/purposesong/view/Register.fxml
@@ -3,6 +3,7 @@
 <?import java.net.URL?>
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Button?>
+<?import javafx.scene.control.CheckBox?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.PasswordField?>
 <?import javafx.scene.control.TextField?>
@@ -53,12 +54,12 @@
                               <Font name="Gill Sans MT" size="30.0" />
                            </font>
                         </Label>
-                        <TextField fx:id="FullNameField" alignment="TOP_LEFT" layoutX="165.0" layoutY="112.0" prefHeight="27.0" prefWidth="458.0" promptText="Type your name..." style="-fx-background-color: black;" styleClass="userName" stylesheets="@../themes/global.css">
+                        <TextField fx:id="UserNameField" alignment="TOP_LEFT" layoutX="165.0" layoutY="112.0" prefHeight="27.0" prefWidth="458.0" promptText="Type your username..." style="-fx-background-color: black;" styleClass="userName" stylesheets="@../themes/global.css">
                            <font>
                               <Font name="Bookshelf Symbol 7" size="12.0" />
                            </font>
                         </TextField>
-                        <Label layoutX="55.0" layoutY="108.0" prefHeight="35.0" prefWidth="105.0" text="FULL NAME:" textFill="WHITE">
+                        <Label layoutX="55.0" layoutY="108.0" prefHeight="35.0" prefWidth="105.0" text="USERNAME:" textFill="WHITE">
                            <font>
                               <Font name="Arial Bold" size="12.0" />
                            </font>
@@ -74,9 +75,10 @@
                               <PerspectiveTransform />
                            </effect>
                         </Button>
-                        <TextField fx:id="EmailField" alignment="TOP_LEFT" layoutX="167.0" layoutY="164.0" prefHeight="27.0" prefWidth="458.0" promptText="Type your name..." style="-fx-background-color: black;" styleClass="userName" stylesheets="@../themes/global.css" />
+                        <TextField fx:id="EmailField" alignment="TOP_LEFT" layoutX="167.0" layoutY="164.0" prefHeight="27.0" prefWidth="458.0" promptText="Type your email..." style="-fx-background-color: black;" styleClass="userName" stylesheets="@../themes/global.css" />
                         <Button layoutX="49.0" layoutY="320.0" mnemonicParsing="false" onAction="#goToMenu" prefHeight="35.0" prefWidth="93.0" style="-fx-background-color: #800a6f;" text="back" />
-                        <Button layoutX="213.0" layoutY="320.0" mnemonicParsing="false" prefHeight="35.0" prefWidth="294.0" style="-fx-background-color: #800a6f;" text="Register" />
+                        <Button layoutX="213.0" layoutY="320.0" mnemonicParsing="false" onAction="#createNewUser" prefHeight="35.0" prefWidth="294.0" style="-fx-background-color: #800a6f;" text="Register" />
+                        <CheckBox fx:id="isVip" layoutX="642.0" layoutY="117.0" mnemonicParsing="false" text="Vip User?" textFill="WHITE" />
                      </children>
                      <opaqueInsets>
                         <Insets />
